+++ History.md
@@ -13,6 +13,7 @@
 * Adds merging adjacent selectors within a scope (single and multiple ones).
 * Changes behavior of `--keep-line-breaks`/`keepBreaks` option to keep breaks after trailing braces only.
 * Makes all multiple selectors ordered alphabetically (aids merging).
+* Adds property overriding so more coarse properties override more granular ones.
 
 1.1.7 / 2013-10-28
 ==================